Probuje uruchomic cmake z qt, Solver do testowania i poprawy

+++ Aplikacja/Solver/include/Functions/FunctionWrapper.hpp
@@ -17,16 +17,16 @@ class FunctionWrapper
 public:
     FunctionWrapper();
 
-    void addSymbols(const std::vector<std::pair<std::string, float>>&);
-    exprtk::expression<float>& configAndGetExpression(const std::string&);
-    std::optional<float> operator ()(const SVector&);
+    void addSymbols(const std::vector<std::pair<std::string, double>>&);
+    exprtk::expression<double>& configAndGetExpression(const std::string&);
+    std::optional<double> operator ()(const SVector&);
     std::string getExpressionString() const;
     unsigned int getDimension() const;
 
 private:
-    std::vector<float> _symbols;
-    exprtk::symbol_table<float> _symbolTable;
-    exprtk::expression<float> _expression;
+    std::vector<double> _symbols;
+    exprtk::symbol_table<double> _symbolTable;
+    exprtk::expression<double> _expression;
     std::string _expressionString;
     Logger& _log;
 };
