+++ Aplikacja/Solver/include/SMatrix.hpp
@@ -14,11 +14,11 @@ class SMatrix
 {
 public:
     SMatrix();
-    explicit SMatrix(const std::vector<std::vector<float>>& vector);
+    explicit SMatrix(const std::vector<std::vector<double>>& vector);
 
-    const std::vector<std::vector<float>>& getVector() const;
-    void setVector(const std::vector<std::vector<float>>& vector);
-    std::optional<float> getMinor(unsigned int&) const;
+    const std::vector<std::vector<double>>& getVector() const;
+    void setVector(const std::vector<std::vector<double>>& vector);
+    std::optional<double> getMinor(unsigned int&) const;
     bool operator==(const SMatrix&)const;
     bool operator!=(const SMatrix&)const;
     const SMatrix& operator -();
@@ -27,17 +27,17 @@ public:
 private:
     /// First-number of rows, Second-number of columns
     std::pair<unsigned int,unsigned int> _dimension;
-    std::vector<std::vector<float>> _vector;
+    std::vector<std::vector<double>> _vector;
     Logger& _log;
 
-    friend SMatrix operator *(const SMatrix&, float);
+    friend SMatrix operator *(const SMatrix&, double);
     friend SMatrix operator *(const SMatrix&, const SMatrix&);
     friend SMatrix operator +(const SMatrix&, const SMatrix&);
     friend SMatrix operator -(const SMatrix&, const SMatrix&);
 
 };
 
-SMatrix operator *(const SMatrix&, float);
+SMatrix operator *(const SMatrix&, double);
 SMatrix operator *(const SMatrix&, const SMatrix&);
 SMatrix operator +(const SMatrix&, const SMatrix&);
 SMatrix operator -(const SMatrix&, const SMatrix&);
