+++ Aplikacja/Solver/sources/Functions/ParserWrapper.cpp
@@ -16,7 +16,7 @@ std::shared_ptr<FunctionWrapper>
 ParserWrapper::parseToFunction(unsigned int dimension, const std::string& expr)
 {
     auto wrapper = std::make_shared<FunctionWrapper>();
-    std::vector<std::pair<std::string,float>> namesAndValues;
+    std::vector<std::pair<std::string,double>> namesAndValues;
     for (auto i = 1; i <= dimension; ++i)
     {
         namesAndValues.emplace_back("x" + std::to_string(i),0.f);
