+++ Aplikacja/Solver/tests/SMatrixTests.cpp
@@ -11,7 +11,7 @@ class SMatrixTests : public ::testing::Test
 public:
     SMatrixTests()
     {
-        std::vector<std::vector<float>> matrix = {{  1, 2, 3}
+        std::vector<std::vector<double>> matrix = {{  1, 2, 3}
                                                   , {9, 2, 9}
                                                   , {5, 2, 0}};
         _sut = std::make_unique<SMatrix>(matrix);
@@ -23,7 +23,7 @@ public:
 
 TEST_F(SMatrixTests, ShouldFindOutDifferenceBetweenMatrices)
 {
-    std::vector<std::vector<float>> matrix = {{  1, 2, 3}
+    std::vector<std::vector<double>> matrix = {{  1, 2, 3}
                                               , {9, 2, 9}
                                               , {5, 2, 1}};
     SMatrix comparingMatrix(matrix);
@@ -34,10 +34,10 @@ TEST_F(SMatrixTests, ShouldFindOutDifferenceBetweenMatrices)
 TEST_F(SMatrixTests, OperationOfDifferencingShouldReturnCorrectResult)
 {
 
-    std::vector<std::vector<float>> substrat = {{  0, 2, 3}
+    std::vector<std::vector<double>> substrat = {{  0, 2, 3}
                                               , {9, 1, 9}
                                               , {5, 2, -1}};
-    std::vector<std::vector<float>> product = {{  1, 0, 0}
+    std::vector<std::vector<double>> product = {{  1, 0, 0}
                                               , {0, 1, 0}
                                               , {0, 0, 1}};
     SMatrix substractMatrix(substrat), productMatrix(product);
@@ -49,7 +49,7 @@ TEST_F(SMatrixTests, OperationOfDifferencingShouldReturnCorrectResult)
 
 TEST_F(SMatrixTests, ShouldCorrectlyMultpiplyMatrices)
 {
-    std::vector<std::vector<float>> substrat = {{  1, 0, 0}
+    std::vector<std::vector<double>> substrat = {{  1, 0, 0}
                                                , {0, 1, 0}
                                                , {0, 0, 1}};
     SMatrix multiplier(substrat);
