+++ Aplikacja/Solver/tests/SVectorTests.cpp
@@ -15,7 +15,7 @@ public:
         _sut = std::make_unique<SVector>();
     }
 
-    void startService(std::vector<float> &vector)
+    void startService(std::vector<double> &vector)
     {
         _sut->setVector(vector);
     }
@@ -25,16 +25,16 @@ public:
 
 TEST_F(SVectorTests, ShouldCorrectlyAddTwoVectors)
 {
-    std::vector<float> first, second;
+    std::vector<double> first, second;
     for (int i = 0; i < 10; ++i)
     {
-        first.emplace_back(static_cast<float>(i));
-        second.emplace_back(static_cast<float>(2 * i));
+        first.emplace_back(static_cast<double>(i));
+        second.emplace_back(static_cast<double>(2 * i));
     }
     startService(first);
     SVector rhs(second);
     auto wynik = (*_sut) + rhs;
-    float check = 0;
+    double check = 0;
     for (const auto& item : wynik.getVector())
     {
         ASSERT_EQ(item, check + (2 * check));
@@ -44,17 +44,17 @@ TEST_F(SVectorTests, ShouldCorrectlyAddTwoVectors)
 
 TEST_F(SVectorTests, ShouldCorrectlyMultiplyVector)
 {
-    std::vector<float> first;
+    std::vector<double> first;
     for (int i = -5; i < 5; ++i)
     {
-        first.emplace_back(static_cast<float>(i));
+        first.emplace_back(static_cast<double>(i));
     }
     startService(first);
 
 
-    float multiplier = 2;
+    double multiplier = 2;
     auto wynik = (*_sut)*multiplier;
-    float check = -5;
+    double check = -5;
     for (const auto& item : wynik.getVector())
     {
         ASSERT_EQ(item, check*multiplier);
@@ -65,15 +65,15 @@ TEST_F(SVectorTests, ShouldCorrectlyMultiplyVector)
 TEST_F(SVectorTests, ShouldCorrectlySubstractTwoVectors)
 {
     unsigned long numberOfElements = 121;
-    std::vector<float> first(numberOfElements, 1.f), second;
+    std::vector<double> first(numberOfElements, 1.f), second;
     startService(first);
     for (int i = 0; i < 10; ++i)
     {
-        second.emplace_back(static_cast<float>(i));
+        second.emplace_back(static_cast<double>(i));
     }
     SVector rhs(second);
     auto wynik = rhs - (*_sut);
-    float check = 0;
+    double check = 0;
     for (const auto& item : wynik.getVector())
     {
         ASSERT_EQ(item, check - 1);
@@ -84,10 +84,10 @@ TEST_F(SVectorTests, ShouldCorrectlySubstractTwoVectors)
 TEST_F(SVectorTests,ShouldComputeCartesianCorrectlyNorm)
 {
     unsigned long numberOfElements = 121;
-    std::vector<float> vector(numberOfElements, 1.f);
+    std::vector<double> vector(numberOfElements, 1.f);
     startService(vector);
 
-    float expectedNorm = 11.f;
+    double expectedNorm = 11.f;
     auto norm = _sut->getCartesianNorm();
     ASSERT_EQ(expectedNorm,norm);
 }
@@ -96,10 +96,10 @@ TEST_F(SVectorTests,ShouldComputeCartesianCorrectlyNorm)
 TEST_F(SVectorTests,ShouldComputeCartesianCorrectlyNormForNegativeElements)
 {
     unsigned long numberOfElements = 121;
-    std::vector<float> vector(numberOfElements, -1.f);
+    std::vector<double> vector(numberOfElements, -1.f);
     startService(vector);
 
-    float expectedNorm = 11.f;
+    double expectedNorm = 11.f;
     auto norm = _sut->getCartesianNorm();
     ASSERT_EQ(expectedNorm,norm);
 }
@@ -107,7 +107,7 @@ TEST_F(SVectorTests,ShouldComputeCartesianCorrectlyNormForNegativeElements)
 TEST_F(SVectorTests, ShouldCorrectlyCheckIfVectorContainsOnlyZeros)
 {
     unsigned long numberOfElements = 121;
-    std::vector<float> vec(numberOfElements, 0);
+    std::vector<double> vec(numberOfElements, 0);
     startService(vec);
     auto containsOnlyZeros = _sut->containsOnlyZeros();
     ASSERT_TRUE(containsOnlyZeros);
@@ -116,7 +116,7 @@ TEST_F(SVectorTests, ShouldCorrectlyCheckIfVectorContainsOnlyZeros)
 TEST_F(SVectorTests, ShouldCorrectlyCheckIfVectorContainsNotOnlyZeros)
 {
     unsigned long numberOfElements = 121;
-    std::vector<float> vec(numberOfElements, 0);
+    std::vector<double> vec(numberOfElements, 0);
     vec[numberOfElements-1] = 1;
     startService(vec);
     auto containsOnlyZeros = _sut->containsOnlyZeros();
@@ -130,7 +130,7 @@ TEST_F(SVectorTests, ShouldCorrectlyCheckIfVectorContainsNotOnlyZeros)
 
 TEST_F(SVectorTests, ShouldReturnNegativeVector)
 {
-    std::vector<float> vec{-1.f,-5.2,2.f,0,111.21,-123.f};
+    std::vector<double> vec{-1.f,-5.2,2.f,0,111.21,-123.f};
     startService(vec);
     auto negative = -(*_sut);
